Terrain Object base class and two terrain objects: slippery floor and slow down floor.
wasd to move. Slow down floor slows you down even if you are in the slip state, which occurs when you go over a slippery tile. In the slippery state, you continue moving in the direction you walked on the tile, but you can't move for a certain amount of time.