New terrain, invulnerability, hp display, movement
The new terrain objects are spikes, walls, and invisible walls.
After getting hurt by spikes, player is invulnerable for a short while.
A short explosion particle effect occurs as a placeholder for knowing we are damaged.
The player's hp is drawn at the top right as text.
Fixed movement, so it's more fluid. The code probably needs to be refactored.
Terrain objects were renamed using the convention obj_terrain_objectname instead of the convention obj_objectname.
Many comments added.