Allow player add items to Inventory
When a player walks over an item, the item is put into the next
available slot in the Inventory.

The new inventory is using instance id's rather than strings to store
items. This is better than using strings because if we want to level
up items we need to store that information into the inventory as well.
In addition, dealing with item cooldowns is likely easier with instance
ids because we don't have to write the cooldown timer again when
equipping the item.