Add init room and toggle inventory visibility
The init room can be used for testing. In it's creation code, we
can call any initialization scripts and point it to the starting room.

The inventory can be toggled by the key global.key_toggle_inventory.
This is purposely abstracted so we can later allow the player to
rebind it to any key they want.

An alternative method is the built in function keyboard_set_map, which
allows us to map one key to another. However, this probably does not
work with mouse buttons, so having an arbitrary variable might be
better if we want to allow the player to rebind mouse buttons as well.