Add item that allows player to fly
Also set the depth of obj_char to -10, so when the player flies,
they are above most things except the inventory.

Currently if the player unequips the item while they are over a wall,
they will get stuck. We can fix this in the future.

TODO:

Allow enemies to equip it.
Prevent player from getting stuck in nonescapable terrain.