Fixes bug with slippery terrain and teleport item
In the randomly generated worlds, slippery terrain would disable character
movement for too long (or indefinetly if the terrain caused the character
to move out of the room).

In the randomly generated worlds, the teleport item could not teleport the
character since the location was not empty. Now, instead of checking for
an empty spot, the item checks for the floor object.