Prevent player from accidentally starting puzzle
The puzzle was starting on mouse release. Now, the player must
completely press the mouse button to start the puzzle. This prevents
the player from accidentally launching the puzzle when entering the
puzzle room during a fight if they were holding down a mouse button.

If the player is constantly clicking the mouse button, this patch
won't help.